Moved favorites to user folder
