+++ src/App.jsx
@@ -12,7 +12,7 @@ import ProductPage from './components/pages/user/productPage/ProductPage'
 import Profile from './components/pages/user/profile/Profile'
 import LandingPage from './components/pages/onBoarding/LandingPage'
 import SigninOptions from './components/pages/onBoarding/SigninOptions'
-import Favourites from './components/pages/favourites/Favourites'
+import Favourites from './components/pages/user/favourites/Favourites'
 import ForgotPassword from './components/pages/login/ForgotPassword'
 import VerifyCode from './components/pages/login/VerifyCode'
 import AlternativesPage from './components/pages/AlternativesPage'
